Adding variable QCOM_TARGET_PRODUCT to refer to board-name
Henceforth, QCOM_TARGET_PRODUCT will be used to refer to board name in
makefiles in place of TARGET_PRODUCT. This is because, TARGET_PRODUCT
is supposed to refer to the end-user visible name for the product.
Hence, our usage of TARGET_PRODUCT as the board name conflicts with the
customers' setting TARGET_PRODUCT to their specific product names.
All occurrences of TARGET_PRODUCT in makefiles should be replaced with
QCOM_TARGET_PRODUCT.

Change-Id: Id587447508b823bfee455ffe417ca78da5234c5a